qweecoco/requirements.txt
Django==1.9
djangorestframework==3.6.2
psycopg2==2.7.1
django-suit==0.2.25
django-filter==1.0.2